fix: 0609 모든 테이블의 pk 데이터 타입 Integer로 통일

+++ src/main/java/com/honbabmap/backend/user/dto/LoginResponse.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 public class LoginResponse {
     private String message;
     private String accessToken;
-    private Long userId; // loginId가 아니라 pk user_id
+    private Integer userId; // loginId가 아니라 pk user_id
     private String nickname;
     private Integer honbabLevel;
 }
